Add text -simdutf to cabal.project
This doesn't change the behavior of the cabal dependency resolution,
but this matches the current state of the environment:
when cabal uses the GHC-bundled build of text, cabal will ignore
the flag settings, and the `simdutf` flag is disabled in that build.

Note that this doesn't cure the checksource-gen issue,
so we still need to configure `-simdutf` in the Main of checksource-gen.